backend/requirements.txt
# 生产环境依赖
# Web框架和API
fastapi[standard]>=0.114.2,<1.0.0
python-multipart>=0.0.7,<1.0.0

# 数据验证和设置
pydantic>2.0
pydantic-settings>=2.2.1,<3.0.0
email-validator>=2.1.0.post1,<3.0.0.0

# 密码和安全
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
pyjwt>=2.8.0,<3.0.0

# 数据库相关
sqlmodel>=0.0.21,<1.0.0
psycopg[binary]>=3.1.13,<4.0.0
alembic>=1.12.1,<2.0.0

# HTTP客户端和工具
httpx>=0.25.1,<1.0.0
tenacity>=8.2.3,<9.0.0

# 邮件和模板
emails>=0.6,<1.0
jinja2>=3.1.4,<4.0.0

# 监控和错误追踪
sentry-sdk[fastapi]>=1.40.6,<2.0.0 